Allow pulling subs title from title metadata.
In subs (subsections?) listing snippet, we currently look for a NodeSeq
value in the `subs` metadatum to populate the title. As an alternative,
we now provide the ability to specify an explicit `title` metadatum,
which will be considered higher precedence.

This allows the title for a markdown document to be specified in YAML,
though note that it isn’t yet ideal (ideally we’d pull it from a
Markdown h1 unless otherwise specified).