Support extracting asciidoc title for page title.
If a doctitle metadatum is not set in the asciidoc metadata, we extract
the first h1 element on the page and assume it is the page
title (which should be a safe assumption). The title metadatum can also
be used to override the doctitle auto-extracted from the level-0
heading in the asciidoc file.